More precise incrementalism
We now require that input queries do not have any dependencies. This
allows us to reliably recompute inputs to check if they have changed
without worrying about the state of dependencies. This allows us to
properly invalidate input queries and recompute them immediately, only
forcing a recomputation in the dependency graph if the input's value
changed.